+++ sdk/src/main/java/ly/count/android/sdk/CountlyConfig.java
@@ -201,6 +201,7 @@ public class CountlyConfig {
 
     // Requests older than this value in hours would be dropped (0 means this feature is disabled)
     int dropAgeHours = 0;
+    String serverConfiguration;
 
     /**
      * THIS VARIABLE SHOULD NOT BE USED
@@ -1008,6 +1009,17 @@ public class CountlyConfig {
     
     protected synchronized CountlyConfig disableHealthCheck() {
         healthCheckEnabled = false;
+        return this;
+    }
+
+    /**
+     * Set the server configuration to be set while initializing the SDK
+     *
+     * @param serverConfiguration The server configuration to be set
+     * @return Returns the same config object for convenient linking
+     */
+    public synchronized CountlyConfig setServerConfiguration(String serverConfiguration) {
+        this.serverConfiguration = serverConfiguration;
         return this;
     }