+++ sdk/src/main/java/ly/count/android/sdk/ModuleContent.java
@@ -11,6 +11,8 @@ import java.util.List;
 import java.util.Locale;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
+import org.jetbrains.annotations.NotNull;
+import org.json.JSONArray;
 import org.json.JSONObject;
 
 public class ModuleContent extends ModuleBase {
@@ -19,7 +21,7 @@ public class ModuleContent extends ModuleBase {
     CountlyTimer countlyTimer;
     private boolean shouldFetchContents = false;
     private boolean isCurrentlyInContentZone = false;
-    private final int zoneTimerInterval;
+    private int zoneTimerInterval;
     private final ContentCallback globalContentCallback;
     private int waitForDelay = 0;
     private final int CONTENT_START_DELAY_MS = 4000; // 4 seconds
@@ -33,6 +35,27 @@ public class ModuleContent extends ModuleBase {
         countlyTimer = new CountlyTimer();
         zoneTimerInterval = config.content.zoneTimerInterval;
         globalContentCallback = config.content.globalContentCallback;
+    }
+
+    @Override
+    void onSdkConfigurationChanged(@NonNull CountlyConfig config) {
+        zoneTimerInterval = config.content.zoneTimerInterval;
+        if (!configProvider.getContentZoneEnabled()) {
+            exitContentZoneInternal();
+        } else {
+            if (!shouldFetchContents) {
+                exitContentZoneInternal();
+            }
+            waitForDelay = 0;
+            enterContentZoneInternal(new String[] {});
+        }
+    }
+
+    @Override
+    void initFinished(@NotNull CountlyConfig config) {
+        if (configProvider.getContentZoneEnabled()) {
+            enterContentZoneInternal(new String[] {});
+        }
     }
 
     void fetchContentsInternal(@NonNull String[] categories) {
@@ -77,16 +100,23 @@ public class ModuleContent extends ModuleBase {
         }, L);
     }
 
-    void registerForContentUpdates(@Nullable String[] categories) {
+    void enterContentZoneInternal(@Nullable String[] categories) {
+        if (!consentProvider.getConsent(Countly.CountlyFeatureNames.content)) {
+            L.w("[ModuleContent] enterContentZoneInternal, Consent is not granted, skipping");
+            return;
+        }
+
+        shouldFetchContents = true;
+
         if (deviceIdProvider.isTemporaryIdEnabled()) {
-            L.w("[ModuleContent] registerForContentUpdates, temporary device ID is enabled, skipping");
+            L.w("[ModuleContent] enterContentZoneInternal, temporary device ID is enabled, skipping");
             return;
         }
 
         String[] validCategories;
 
         if (categories == null) {
-            L.w("[ModuleContent] registerForContentUpdates, categories is null, providing empty array");
+            L.w("[ModuleContent] enterContentZoneInternal, categories is null, providing empty array");
             validCategories = new String[] {};
         } else {
             validCategories = categories;
@@ -100,20 +130,21 @@ public class ModuleContent extends ModuleBase {
             contentInitialDelay = CONTENT_START_DELAY_MS;
         }
 
-        countlyTimer.startTimer(zoneTimerInterval, contentInitialDelay, () -> {
-            L.d("[ModuleContent] registerForContentUpdates, waitForDelay: [" + waitForDelay + "], shouldFetchContents: [" + shouldFetchContents + "], categories: [" + Arrays.toString(validCategories) + "]");
+        countlyTimer.startTimer(zoneTimerInterval, contentInitialDelay, new Runnable() {
+            @Override public void run() {
+                L.d("[ModuleContent] registerForContentUpdates, waitForDelay: [" + waitForDelay + "], shouldFetchContents: [" + shouldFetchContents + "], categories: [" + Arrays.toString(validCategories) + "]");
+                if (waitForDelay > 0) {
+                    waitForDelay--;
+                    return;
+                }
 
-            if (waitForDelay > 0) {
-                waitForDelay--;
-                return;
+                if (!shouldFetchContents) {
+                    L.w("[ModuleContent] enterContentZoneInternal, shouldFetchContents is false, skipping");
+                    return;
+                }
+
+                fetchContentsInternal(validCategories);
             }
-
-            if (!shouldFetchContents) {
-                L.w("[ModuleContent] registerForContentUpdates, shouldFetchContents is false, skipping");
-                return;
-            }
-
-            fetchContentsInternal(validCategories);
         }, L);
     }