+++ sdk/src/main/java/ly/count/android/sdk/ModuleCrash.java
@@ -306,7 +306,7 @@ public class ModuleCrash extends ModuleBase {
         }
 
         if (!configProvider.getCrashReportingEnabled()) {
-            L.w("[ModuleCrash] recordExceptionInternal, Crash reporting is disabled in the server configuration");
+            L.d("[ModuleCrash] recordExceptionInternal, Crash reporting is disabled in the server configuration");
             return _cly;
         }
 
@@ -342,6 +342,11 @@ public class ModuleCrash extends ModuleBase {
 
     @Override
     void initFinished(@NonNull CountlyConfig config) {
+        if (!configProvider.getCrashReportingEnabled()) {
+            L.d("[ModuleCrash] initFinished, Crash reporting is disabled in the server configuration");
+            return;
+        }
+
         //enable unhandled crash reporting
         if (!configProvider.getCrashReportingEnabled()) {
             L.w("[ModuleCrash] initFinished, Crash reporting is disabled in the server configuration");